requirements.txt
streamlit
nbformat
torch
transformers
faiss-cpu
langchain
python-dotenv
pickle
